feat: add migration details format and refactor in form and details in separated components
